refactored utils.go to vmmanager and added StoppedTime API
AcceptConnection method was incorrectly part of guestmanager package. Moreover, we needed stop time in SandboxWait API and hence the same is added to the manager.
